% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.picked.R
\name{count.picked}
\alias{count.picked}
\title{Compare haplogroup counts in observed population and resampled result}
\usage{
count.picked(observed, picked)
}
\arguments{
\item{observed}{A vector of observed absolute frequencies to be compared to the ones produced by repeated resampling by the 'pick' function. This is typically a single column from the input file, i.e., the observed numbers of a set of haplogroups in a (small) sample from a (ancient) population.}

\item{picked}{A matrix of resampled observations produced by the 'pick' function.}
}
\value{
A numeric vector of the number of permutations equal to or more extreme than the observed data.
}
\description{
Count, from a matrix produced by the 'pick' function, the number of permutations that produced a result equal to or more extreme than the one in the observed population.
}
