% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Resample a reference population (to a smaller size) and count the number of haplogroup observations}
\usage{
pick(freqs, n2pick, rounds, zerofreq = 0)
}
\arguments{
\item{freqs}{The vector of relative frequencies observed in the reference population.}

\item{n2pick}{The number of observations to pick from the reference (with replacement).}

\item{rounds}{The number of times the picking is to be repeated.}

\item{zerofreq}{The frequency to be used in the picking for haplogroups (which are observed in the sample populations and) whose observed frequency in the reference population is zero. Defaults to 0.}
}
\value{
A matrix of the numbers of times that each haplogroup was observed in each resampling.
}
\description{
Pick n2pick items with each element of freqs as the probability of success in turn. Repeat the picking rounds times. For freqs\link{i}==0, substitute the probability of success by zerofreq.
In practice, this answers the question of whether sampling error alone could explain the observed frequency differences; freqs come from the reference population and n2pick is the sample size of the (ancient) population of interest.
The parameter zerofreq acknowledges that the reference freqs may be inaccurate; however, all nonzero freqs are assumed correct - i.e., for small reference populations, use a more elegant procedure.
}
