% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.picked.R
\name{plot.picked}
\alias{plot.picked}
\title{Plot, from a matrix produced by the pick function, the empirical distributions and the real observation}
\usage{
\method{plot}{picked}(
  picked,
  tot_n,
  obs_n,
  text = 1:nrow(picked),
  probl = NULL,
  plot = T,
  histcol = "gray",
  obscol = "red",
  lineheight = 0.8,
  textoffsety = 0.3,
  main = ""
)
}
\arguments{
\item{picked}{The input matrix, produced by the 'pick' function.}

\item{tot_n}{The total number of observations in the sampled population.}

\item{obs_n}{The vector of observations in the sampled population.}

\item{probl}{A logical vector indicating which observations differ significantly from expected; their haplogroup names will be plotted in red in the histograms.}

\item{plot}{Should the histograms be plotted? Defaults to TRUE. If FALSE, returns the histogram values that could be plotted by hist.}

\item{histcol}{Plotting colour of the histogram bars. Defaults to gray.}

\item{obscol}{Plotting colour of the vertical bars for the true observed frequencies. Defaults to red.}

\item{lineheight}{A plotting parameter for the histograms. Defaults to 0.8.}

\item{textoffsety}{A plotting parameter for the histograms. Defaults to 0.3.}

\item{main}{A main title for the plot. Defaults to no title.}

\item{text=1:nrow(picked)}{The vector of haplogroup names. Defaults to a running number.}
}
\value{
A set of histogram plots (if plot==T) or a set of values that could be used to plot those histograms (if plot==F).
}
\description{
This function plots a histogram of empirical values produced by the 'pick' function and depicts the observed frequency from the sampled (ancient) population of interest by a vertical bar.
}
