% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.continuity.R
\name{test.continuity}
\alias{test.continuity}
\title{Test whether the haplogroup distributions in small samples are compatible with that in a larger reference sample}
\usage{
test.continuity(freq, perms = 10000, signf = 0.05, plotTF = T)
}
\arguments{
\item{freq}{Input matrix of haplogroup observations. 1st row is a header with population names (these will be used as PDF plot names). Each remaining row corresponds to one haplogroup. 1st column has haplogroup names (to be used as histogram titles on the plots). 2nd column has haplogroup observations (absolute frequencies) from a reference population, and from 3rd column onwards each column has the corresponding observations for other (smaller, potentially ancient) populations.}

\item{perms}{How many times should the resampling be repeated (with replacement). Defaults to 10000.}

\item{signf}{The nominal significance to use for marking the haplogroups that differ from expected. Defaults to 0.05. This is treated as two-tailed, so threshold is half of the given probability value at either end of the distribution.}

\item{plotTF}{Should the resulting plots be plotted? Defaults to TRUE.}
}
\value{
A set of PDF plots in the active working directory (if plotTF==T, the default).
}
\description{
This function resamples a reference population 'perms' times, and compares the resulting empirical distribution to the observed number of haplogroups in one or more (smaller) samples from other (potentially ancient) populations. The result for each tested population is plotted into one PDF figure, with histograms showing the empirical distributions based on the resampling, vertical bars depicting the observed haplogroup frequencies in the tested population, and red colour of haplogroup names indicating those whose frequency differs nominally significantly from the expectation.
}
